requirements.txt
requests
numpy
sentence_transformers
transformers
scipy
torch
peft
datasets
PyYAML
jsonlines
python-dotenv
accelerate